Agents_Experience/requirements.txt
# Agent交易系统依赖

# HTTP请求
requests>=2.31.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理（继承自主项目）
pandas>=2.0.0
numpy>=1.24.0

# 可选：OpenAI SDK (如果使用标准OpenAI格式API)
# openai>=1.0.0
